refactor: split twitter-client
